# optimized `lastLine()` usage in `readfile()`