@firewave 2023-09-30
The remaining `lastLine()` can also be removed but that gets rather messy.